multiselect: Dynamic thumbnails for buildings
Can still look funny sometimes due to scaling issues, and can also
change the thumbnail of a group (2+ buildings of same class but with
different graphics) when adding or removing instances from
multiselection. That bug also affects positioning of instance groups in
the multiselection tab and is not caused by the changes in this commit.